Fix Int to Long cast for route id in delete event

+++ app/src/main/kotlin/biz/cesena/packride4/ui/savedroutes/SavedRoutesViewModel.kt
@@ -22,7 +22,7 @@ class SavedRoutesViewModel @Inject constructor(
     fun delete(route: SavedRoute) {
         viewModelScope.launch {
             dao.delete(route)
-            routeEventBus.notifyRouteDeleted(route.id)
+            routeEventBus.notifyRouteDeleted(route.id.toLong())
         }
     }
 }
